requirements.txt
github3.py==0.9.4
requests==2.8.1
uritemplate.py==0.3.0
urllib3==1.12
